Refactor typst-render
